Fixing tests and other WTFs
