+++ tests/src/Functional/MessageUiMassiveHardCodedArgumentsTest.php
@@ -17,23 +17,7 @@ use Drupal\user\UserInterface;
  *
  * @group Message UI
  */
-class MessageUiMassiveHardCodedArgumentsTest extends MessageTestBase {
-
-  /**
-   * The user object.
-   *
-   * @var UserInterface
-   */
-  public $user;
-
-  /**
-   * Modules to enable.
-   *
-   * @todo: verify whether entity_token is required in D8.
-   *
-   * @var array
-   */
-  public static $modules = ['message', 'message_ui'];
+class MessageUiMassiveHardCodedArgumentsTest extends AbstractTestMessageUi {
 
   /**
    * @var MessageTemplateInterface
@@ -46,7 +30,7 @@ class MessageUiMassiveHardCodedArgumentsTest extends MessageTestBase {
   public function setUp() {
     parent::setUp();
 
-    $this->user = $this->drupalCreateUser();
+    $this->account = $this->drupalCreateUser();
   }
 
   /**
@@ -58,20 +42,20 @@ class MessageUiMassiveHardCodedArgumentsTest extends MessageTestBase {
 
     // Set a queue worker for the update arguments when updating a message
     // template.
-    $this->configSet('update_tokens.update_tokens', TRUE, 'message_ui.settings');
-    $this->configSet('update_tokens.how_to_update', 'update_with_item', 'message_ui.settings');
+    $this->configSet('update_tokens.update_tokens', TRUE);
+    $this->configSet('update_tokens.how_to_update', 'update_with_item');
 
     /* @var $message Message */
     $message = Message::create(['template' => $this->messageTemplate->id()]);
 
     $message
-      ->setOwner($this->user)
+      ->setOwner($this->account)
       ->save();
 
     $original_arguments = $message->getArguments();
 
     // Update message instance when removing a hard coded argument.
-    $this->configSet('update_tokens.how_to_act', 'update_when_removed', 'message_ui.settings');
+    $this->configSet('update_tokens.how_to_act', 'update_when_removed');
 
     // Set message text.
     $this->messageTemplate->set('text', [
@@ -91,19 +75,17 @@ class MessageUiMassiveHardCodedArgumentsTest extends MessageTestBase {
 
     // Verify the arguments has changed.
     $message = Message::load($message->id());
+
     $this->assertTrue($original_arguments != $message->getArguments(), 'The message arguments has changed during the queue worker work.');
 
     // Creating a new message and her hard coded arguments.
     $message = Message::create(['template' => $this->messageTemplate->id()]);
+    $message->setOwner($this->account)->save();
 
-    $message
-      ->setOwner($this->user)
-      ->save();
     $original_arguments = $message->getArguments();
 
-    // @todo : validate / fix this config access.
     // Process the message instance when adding hard coded arguments.
-    $this->configSet('update_tokens.how_to_act', 'update_when_added', 'message_ui.settings');
+    $this->configSet('update_tokens.how_to_act', 'update_when_added');
 
     $message_template = $this->loadMessageTemplate('dummy_message');
     $message_template->set('text', array('@{message:user:name}.'));
@@ -113,8 +95,6 @@ class MessageUiMassiveHardCodedArgumentsTest extends MessageTestBase {
     $queue = \Drupal::queue('message_ui_arguments');
     $item = $queue->claimItem();
 
-    // @todo: item->data currently returning null, do we need a check for that?
-    // @todo : check the below calls MessageUiArgumentsWorker::createItem.
     $queue->createItem($item->data);
 
     // Verify the arguments has changed.
