+++ tests/src/Functional/MessageUiShowPreviewTest.php
@@ -7,48 +7,15 @@
 
 namespace Drupal\Tests\message_ui\Functional;
 
-use Drupal\Tests\message\Functional\MessageTestBase;
 use Drupal\user\Entity\Role;
 use Drupal\user\RoleInterface;
-use Drupal\user\UserInterface;
 
 /**
  * Testing the display of the preview.
  *
  * @group Message UI
  */
-class MessageUiShowPreviewTest extends MessageTestBase {
-
-  /**
-   * The user account object.
-   *
-   * @var UserInterface
-   */
-  protected $account;
-
-  /**
-   * Modules to enable.
-   *
-   * @var array
-   */
-  public static $modules = ['message', 'message_ui'];
-
-  /**
-   * The user role.
-   *
-   * @var integer
-   */
-  protected $rid;
-
-  /**
-   * Grant to the user a specific permission.
-   *
-   * @param string $operation
-   *   The template of operation - create, update, delete or view.
-   */
-  protected function grantMessageUiPermission($operation) {
-    user_role_grant_permissions($this->rid, array($operation . ' foo message'));
-  }
+class MessageUiShowPreviewTest extends AbstractTestMessageUi {
 
   /**
    * {@inheritdoc}
@@ -69,20 +36,6 @@ class MessageUiShowPreviewTest extends MessageTestBase {
 
     // Don't show the text of the message.
     $this->configSet('show_preview', TRUE);
-  }
-
-  /**
-   * Set a config value.
-   *
-   * @param string $config
-   *   The config name.
-   * @param string $value
-   *   The config value.
-   * @param string $storage
-   *   The storing of the configuration. Default to message.message.
-   */
-  protected function configSet($config, $value, $storage = 'message_ui.settings') {
-    $this->container->get('config.factory')->getEditable($storage)->set($config, $value)->save();
   }
 
   /**
